Change progress type to Double

+++ app/src/main/java/com/vipulog/ebookreader/ReaderActivity.kt
@@ -148,7 +148,7 @@ class ReaderActivity : AppCompatActivity(), EbookReaderEventListener, ActionMode
     }
 
 
-    override fun onProgressChanged(progress: Float, currentTocItem: TocItem?) {
+    override fun onProgressChanged(progress: Double, currentTocItem: TocItem?) {
         this.currentTocItem = currentTocItem
         binding.progressBar.progress = (progress * 100).toInt()
         binding.appBar.title = currentTocItem?.label ?: ""
